Adds datatype conversion to get_player_by_name
Dataframes converted using the dataclass model annotations. The test_get_player_by_name_accents tests were failing because points_per_game was set to int. That is also fixed in this commit.

+++ models/players.py
@@ -39,7 +39,7 @@ class FieldPlayer(PlayerBase):
     goals_pens: int
     pens_made: int
     pens_att: int
-    points_per_game: int
+    points_per_game: float
     on_goals_for: int
     on_goals_against: int
     goals_per90: float
